+++ external/ImageMagick/coders/wmf.c
@@ -17,7 +17,7 @@
 %                               December 2000                                 %
 %                                                                             %
 %                                                                             %
-%  Copyright 1999-2020 ImageMagick Studio LLC, a non-profit organization      %
+%  Copyright 1999-2021 ImageMagick Studio LLC, a non-profit organization      %
 %  dedicated to making software imaging solutions freely available.           %
 %                                                                             %
 %  You may not use this file except in compliance with the License.  You may  %
@@ -839,7 +839,7 @@ static void ipa_device_begin(wmfAPI * API)
   DrawSetViewbox(WmfDrawingWand,0,0,ddata->image->columns,ddata->image->rows);
 
   (void) FormatLocaleString(comment,MagickPathExtent,"Created by %s",
-    GetMagickVersion((unsigned long *) NULL));
+    GetMagickVersion((size_t *) NULL));
   DrawComment(WmfDrawingWand,comment);
 
   /* Scale width and height to image */
